Still needs some work

+++ src/codegen/llvm/mod.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap};
+use std::collections::HashMap;
 
 use inkwell::{
     AddressSpace,
@@ -7,10 +7,10 @@ use inkwell::{
     context::Context,
     module::{Linkage, Module},
     targets::{TargetMachineOptions, TargetTriple},
-    types::{BasicMetadataTypeEnum, BasicTypeEnum, FunctionType},
+    types::{BasicMetadataTypeEnum, BasicTypeEnum, FunctionType, StructType},
     values::{
         BasicMetadataValueEnum, BasicValue, BasicValueEnum, FloatValue, FunctionValue, PhiValue,
-        ValueKind,
+        PointerValue, ValueKind,
     },
 };
 use inkwell::{FloatPredicate, IntPredicate};
@@ -29,8 +29,8 @@ use inkwell::{
 use std::path::Path;
 
 use std::env;
-use std::process::Command;
 use std::fs;
+use std::process::Command;
 pub static FILE_EXTENSION_EXE: &str = if cfg!(target_os = "windows") {
     ".exe"
 } else {
@@ -45,6 +45,8 @@ pub struct LlvmGenerator<'a> {
     curr_func: Option<FunctionValue<'a>>,
     curr_tir_func: Option<Function>,
     phi_fixups: Vec<(PhiValue<'a>, String, BlockId, SSAValue)>,
+    ///Maps TirType Interface -> (LLVM Struct Type, INTERFACE name)
+    struct_interfaces: HashMap<TirType, (StructType<'a>, String)>,
 }
 impl<'a> LlvmGenerator<'a> {
     pub fn new(ctx: &'a Context, main_module: Module<'a>) -> LlvmGenerator<'a> {
@@ -57,6 +59,7 @@ impl<'a> LlvmGenerator<'a> {
             curr_func: None,
             curr_tir_func: None,
             phi_fixups: vec![],
+            struct_interfaces: HashMap::new(),
         };
     }
     fn get_ssa_val(&self, func_name: &str, ssa: SSAValue) -> BasicValueEnum<'a> {
@@ -108,12 +111,11 @@ impl<'a> LlvmGenerator<'a> {
                 SSAValue {
                     val: id,
                     ty: Some(ty),
-                }
+                },
             )),
             TIR::ItoF(id, ssa_ref, ty) => {
                 let ins: FloatValue<'_> = builder.build_signed_int_to_float(
-                    self.get_ssa_val(&curr_func_name, ssa_ref)
-                        .into_int_value(),
+                    self.get_ssa_val(&curr_func_name, ssa_ref).into_int_value(),
                     self.ctx.f64_type(),
                     "sitofp",
                 )?;
@@ -200,12 +202,16 @@ impl<'a> LlvmGenerator<'a> {
                         TirType::I64 => self.ctx.i64_type().fn_type(&compiled_types, false),
                         TirType::F64 => self.ctx.f64_type().fn_type(&compiled_types, false),
                         TirType::I1 => self.ctx.bool_type().fn_type(&compiled_types, false),
-                        TirType::I8PTR => self.ctx.ptr_type(AddressSpace::default()).fn_type(&compiled_types, false),
+                        TirType::I8PTR => self
+                            .ctx
+                            .ptr_type(AddressSpace::default())
+                            .fn_type(&compiled_types, false),
                         TirType::Void => self.ctx.void_type().fn_type(&compiled_types, false),
                         _ => todo!("Chase you have not implemented this return type yet"),
                     };
-                    
-                    self.main_module.add_function(&*name, fn_type, Some(Linkage::External))
+
+                    self.main_module
+                        .add_function(&*name, fn_type, Some(Linkage::External))
                 };
                 let param_types = func_body.get_type().get_param_types();
                 let llvm_params: Vec<BasicMetadataValueEnum> = params
@@ -481,7 +487,8 @@ impl<'a> LlvmGenerator<'a> {
                 ))
             }
             TIR::Not(id, v) => {
-                let val = self.get_ssa_val(&curr_func_name, v.clone())
+                let val = self
+                    .get_ssa_val(&curr_func_name, v.clone())
                     .into_int_value();
 
                 let one = self.ctx.bool_type().const_int(1, false);
@@ -512,10 +519,7 @@ impl<'a> LlvmGenerator<'a> {
                 let first_val_ssa = &vals[0];
                 let mut ty = None;
                 for val in &vals {
-                    if let Some(v) = self
-                        .tir_to_val
-                        .get(&(curr_func_name.clone(), val.clone()))
-                    {
+                    if let Some(v) = self.tir_to_val.get(&(curr_func_name.clone(), val.clone())) {
                         ty = Some(v.get_type());
                         break;
                     }
@@ -552,13 +556,15 @@ impl<'a> LlvmGenerator<'a> {
                                     val: ssa_val.val,
                                     ty: Some(at),
                                 };
-                                if let Some(val) = self.tir_to_val.get(&(curr_func_name.clone(), alt_ssa)) {
+                                if let Some(val) =
+                                    self.tir_to_val.get(&(curr_func_name.clone(), alt_ssa))
+                                {
                                     phi.add_incoming(&[(&*val, *block)]);
                                     found = true;
                                 }
                             }
                         }
-                        
+
                         if !found {
                             self.phi_fixups.push((
                                 phi,
@@ -579,14 +585,21 @@ impl<'a> LlvmGenerator<'a> {
                 ))
             }
             TIR::GlobalString(id, va) => {
-                let str_val = builder.build_global_string_ptr(&va, format!("global_str_{}", id).as_str())?;
+                let str_val =
+                    builder.build_global_string_ptr(&va, format!("global_str_{}", id).as_str())?;
                 let str_ptr = str_val.as_pointer_value();
-                let ptr_i64 = builder.build_ptr_to_int(
-                    str_ptr,
-                    self.ctx.i64_type(),
-                    "str_ptr_to_i64",
-                )?;
-                self.tir_to_val.insert((curr_func_name.clone(), SSAValue { val: id, ty: Some(TirType::I64) }), ptr_i64.into());
+                let ptr_i64 =
+                    builder.build_ptr_to_int(str_ptr, self.ctx.i64_type(), "str_ptr_to_i64")?;
+                self.tir_to_val.insert(
+                    (
+                        curr_func_name.clone(),
+                        SSAValue {
+                            val: id,
+                            ty: Some(TirType::I64),
+                        },
+                    ),
+                    ptr_i64.into(),
+                );
                 Some((
                     ptr_i64.into(),
                     SSAValue {
@@ -595,7 +608,159 @@ impl<'a> LlvmGenerator<'a> {
                     },
                 ))
             }
-            _ => todo!("Chase you have not implemented {:?} ins yet", inst),
+            TIR::CreateStructInterface(_, boxed_name, ty) => {
+                let types = match ty.clone() {
+                    TirType::StructInterface(typ) => typ,
+                    _ => unreachable!(),
+                };
+
+                let custom_struct = self.ctx.opaque_struct_type(&*boxed_name);
+                let mut llvm_types = vec![];
+                for ty in types {
+                    llvm_types.push(self._tir_to_llvm_type(ty));
+                }
+                custom_struct.set_body(&llvm_types, false);
+                self.struct_interfaces
+                    .insert(ty, (custom_struct, *boxed_name));
+
+                None
+            }
+            TIR::CreateStructLiteral(id, ty, vals) => {
+                let (struct_type, interface_name) = self
+                    .struct_interfaces
+                    .get(&ty)
+                    .ok_or_else(|| ToyError::new(ToyErrorType::MissingInstruction))?;
+
+                let allocated_struct = builder.build_alloca(*struct_type, interface_name)?;
+
+                let zero = self.ctx.i32_type().const_int(0, false);
+
+                for (i, val) in vals.iter().enumerate() {
+                    let idx = self.ctx.i32_type().const_int(i as u64, false);
+
+                    let field_ptr = unsafe {
+                        builder.build_gep(
+                            *struct_type,
+                            allocated_struct,
+                            &[zero, idx],
+                            &format!("field_{i}_ptr"),
+                        )?
+                    };
+
+                    builder.build_store(
+                        field_ptr,
+                        self.get_ssa_val(&curr_func_name, val.to_owned()),
+                    )?;
+                }
+                let val_ptr: BasicValueEnum<'_> = allocated_struct.into();
+                self.tir_to_val.insert(
+                    (
+                        curr_func_name.clone(),
+                        SSAValue {
+                            val: id,
+                            ty: Some(ty.clone()),
+                        },
+                    ),
+                    val_ptr,
+                );
+
+                Some((
+                    val_ptr,
+                    SSAValue {
+                        val: id,
+                        ty: Some(ty),
+                    },
+                ))
+            }
+            TIR::ReadStructLiteral(id, struct_ptr, idx) => {
+                let struct_tir_type = struct_ptr
+                    .clone()
+                    .ty
+                    .ok_or_else(|| return ToyError::new(ToyErrorType::UndefinedStruct))?;
+                let field_type = if let TirType::StructInterface(field_types) = &struct_tir_type {
+                    field_types
+                        .get(idx as usize)
+                        .cloned()
+                        .ok_or_else(|| ToyError::new(ToyErrorType::UndefinedStruct))?
+                } else {
+                    return Err(ToyError::new(ToyErrorType::UndefinedStruct));
+                };
+                let (struct_type, _interface_name) =
+                    self.struct_interfaces
+                        .get(&struct_tir_type)
+                        .ok_or_else(|| return ToyError::new(ToyErrorType::UndefinedStruct))?;
+                let struct_literal = self.get_ssa_val(&curr_func_name, struct_ptr.clone());
+                let zero = self.ctx.i32_type().const_int(0u64, false);
+                let compiled_idx = self.ctx.i32_type().const_int(idx, false);
+                let struct_literal_as_struct: PointerValue = match struct_literal {
+                    BasicValueEnum::PointerValue(p) => p,
+                    _ => panic!("Got {:?}", struct_literal.get_type()),
+                };
+                let field_ptr = unsafe {
+                    builder.build_gep(
+                        *struct_type,
+                        struct_literal_as_struct,
+                        &[zero, compiled_idx],
+                        &format!("field_{idx}_ptr"),
+                    )?
+                };
+                let field_val = builder.build_load(
+                    struct_type
+                        .get_field_type_at_index(idx as u32)
+                        .ok_or_else(|| ToyError::new(ToyErrorType::UndefinedStruct))?,
+                    field_ptr,
+                    &format!("field_{idx}_val"),
+                )?;
+                Some((
+                    field_val,
+                    SSAValue {
+                        val: id,
+                        ty: Some(field_type),
+                    },
+                ))
+            }
+            TIR::WriteStructLiteral(id, struct_ptr, field_idx, new_val) => {
+                let struct_tir_type = struct_ptr
+                    .clone()
+                    .ty
+                    .ok_or_else(|| return ToyError::new(ToyErrorType::UndefinedStruct))?;
+                let field_type = if let TirType::StructInterface(field_types) = &struct_tir_type {
+                    field_types
+                        .get(field_idx as usize)
+                        .cloned()
+                        .ok_or_else(|| ToyError::new(ToyErrorType::UndefinedStruct))?
+                } else {
+                    return Err(ToyError::new(ToyErrorType::UndefinedStruct));
+                };
+                let (struct_type, _interface_name) =
+                    self.struct_interfaces
+                        .get(&struct_tir_type)
+                        .ok_or_else(|| return ToyError::new(ToyErrorType::UndefinedStruct))?;
+                let struct_literal = self.get_ssa_val(&curr_func_name, struct_ptr.clone());
+                let value_to_store = self.get_ssa_val(&curr_func_name, new_val.clone());
+                let zero = self.ctx.i32_type().const_int(0u64, false);
+                let compiled_idx = self.ctx.i32_type().const_int(field_idx, false);
+                let struct_literal_as_struct: PointerValue = match struct_literal {
+                    BasicValueEnum::PointerValue(p) => p,
+                    _ => unreachable!(),
+                };
+                let field_ptr = unsafe {
+                    builder.build_gep(
+                        *struct_type,
+                        struct_literal_as_struct,
+                        &[zero, compiled_idx],
+                        &format!("field_{field_idx}_ptr"),
+                    )?
+                };
+                builder.build_store(field_ptr, value_to_store)?;
+                Some((
+                    value_to_store,
+                    SSAValue {
+                        val: id,
+                        ty: Some(field_type),
+                    },
+                ))
+            }
         };
         if let Some((llvm_ir, val)) = res {
             self.tir_to_val.insert((curr_func_name, val), llvm_ir);
@@ -628,7 +793,11 @@ impl<'a> LlvmGenerator<'a> {
                     TirType::F64 => self.ctx.f64_type().into(),
                     TirType::I1 => self.ctx.bool_type().into(),
                     TirType::I8PTR => self.ctx.ptr_type(AddressSpace::default()).into(),
-                    _ => todo!("Chase you have not implemented this param type yet"),
+                    TirType::StructInterface(_) => {
+                        // Structs are passed as pointers for consistency with CreateStructLiteral
+                        self.ctx.ptr_type(AddressSpace::default()).into()
+                    }
+                    _ => todo!("Chase you have not implemented {t:?} param type yet"),
                 },
                 None => unreachable!(), //SAFETY: Guaranteed by ast-gen
             })
@@ -641,7 +810,13 @@ impl<'a> LlvmGenerator<'a> {
                 .ctx
                 .ptr_type(AddressSpace::default())
                 .fn_type(llvm_params.as_slice(), false),
-            _ => todo!("Chase you have not implemented this return type yet"),
+            TirType::StructInterface(_) => {
+                // Structs are returned as pointers for consistency
+                self.ctx
+                    .ptr_type(AddressSpace::default())
+                    .fn_type(llvm_params.as_slice(), false)
+            }
+            TirType::Void => self.ctx.void_type().fn_type(llvm_params.as_slice(), false),
         };
 
         let llvm_func = if let Some(f) = self.main_module.get_function(&func.name) {
@@ -664,13 +839,20 @@ impl<'a> LlvmGenerator<'a> {
         //</Boiler plate to setup function>
         for (n, p) in func.params.iter().enumerate() {
             let p_val = llvm_func.get_nth_param(n as u32).unwrap(); //is probably safe, maybe will cause bugs :D
-            self.tir_to_val
-                .insert((*self.curr_tir_func.as_ref().unwrap().name.clone(), p.clone()), p_val);
+            self.tir_to_val.insert(
+                (
+                    *self.curr_tir_func.as_ref().unwrap().name.clone(),
+                    p.clone(),
+                ),
+                p_val,
+            );
         }
-        
+
         for b in &func.body {
-             let llvm_block = self.ctx.append_basic_block(llvm_func, &format!("block_{}", b.id));
-             self.block_id_to_block.insert(b.id, llvm_block);
+            let llvm_block = self
+                .ctx
+                .append_basic_block(llvm_func, &format!("block_{}", b.id));
+            self.block_id_to_block.insert(b.id, llvm_block);
         }
 
         for b in &func.body {
@@ -693,7 +875,11 @@ impl<'a> LlvmGenerator<'a> {
             TirType::F64 => self.ctx.f64_type().into(),
             TirType::I1 => self.ctx.bool_type().into(),
             TirType::I8PTR => self.ctx.ptr_type(AddressSpace::default()).into(),
-            _ => todo!("Chase you have not implemented this param type yet"),
+            TirType::StructInterface(_) => {
+                // Structs are passed as pointers for consistency
+                self.ctx.ptr_type(AddressSpace::default()).into()
+            }
+            _ => todo!("Chase you have not implemented {t:?} param type yet"),
         };
     }
     fn declare_individual_function(&mut self, name: &str, types: Vec<TirType>, ret_type: TirType) {
@@ -806,6 +992,12 @@ impl<'a> LlvmGenerator<'a> {
             vec![TirType::I64, TirType::I64],
             TirType::I64,
         );
+        self.declare_individual_function(
+            "toy_free",
+            vec![TirType::I64], //?
+            TirType::Void,
+        );
+        self.declare_individual_function("toy_free_arr", vec![TirType::I64], TirType::Void);
         return Ok(());
     }
     fn generate_internal(&mut self, funcs: Vec<Function>) -> Result<(), ToyError> {
@@ -903,7 +1095,6 @@ impl<'a> LlvmGenerator<'a> {
         let rstatus = Command::new(lib_path.join("ld.lld"))
             .args(args.clone())
             .status();
-        println!("RSTATUS: {:#?}, args {:#?}", rstatus, args);
         let status = match rstatus {
             Ok(f) => f,
             Err(_) => return Err(ToyError::new(ToyErrorType::InternalLinkerFailure)),
@@ -912,13 +1103,17 @@ impl<'a> LlvmGenerator<'a> {
             return Err(ToyError::new(ToyErrorType::InternalLinkerFailure));
         }
         let p_args: Vec<String> = env::args().collect();
-        if p_args.clone().contains(&"--repl".to_owned()) || p_args.clone().contains(&"--run".to_owned()) {
+        if p_args.clone().contains(&"--repl".to_owned())
+            || p_args.clone().contains(&"--run".to_owned())
+        {
             let mut prgm = Command::new(format!("{}{}", "./", output_name.as_str()));
             let _ = prgm.spawn().unwrap().wait().unwrap();
 
-            fs::remove_file(output_name.as_str()).unwrap();
-            fs::remove_file(format!("{}.o", prgm_name)).unwrap();
-            fs::remove_file(format!("{}.ll", prgm_name)).unwrap();
+            if !p_args.contains(&"--save-temps".to_string()) {
+                fs::remove_file(output_name.as_str()).unwrap();
+                fs::remove_file(format!("{}.o", prgm_name)).unwrap();
+                fs::remove_file(format!("{}.ll", prgm_name)).unwrap();
+            }
         }
         return Ok(());
     }
